Menu

+++ resources/views/users/index.blade.php
@@ -8,7 +8,9 @@
             <h2>Users Management</h2>
         </div>
         <div class="pull-right">
+            @can('role-create')
             <a class="btn btn-success" href="{{ route('users.create') }}"> Create New User</a>
+            @endcan
         </div>
     </div>
 </div>
@@ -41,12 +43,17 @@
         @endforeach
       @endif
     </td>
+
     <td>
        <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
+       @can('role-edit')
        <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+       @endcan
+       @can('role-delete')
         {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
             {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
         {!! Form::close() !!}
+        @endcan
     </td>
   </tr>
  @endforeach
